Accumulate GPU histograms in Int64 fixed point
Quantise each gradient once per tree with a single power-of-two scale
chosen so nobs * max|∇| * scale <= 2^52, and accumulate the histogram
with Int64 atomics. Sibling subtraction is exact, the result no longer
depends on atomic ordering, and every bin and node total converts back
to Float64 without rounding. A positive hessian that truncates to zero
is floored to 1, following XGBoost's GradientQuantiser.

The split scan runs on the raw quantised values: gains are homogeneous of
degree one in the stats, so L2, min_weight and gamma are scaled alongside
and nodes_sum / tree_gain are divided back once at the end of grow_tree!.

+++ ext/EvoTreesKernelAbstractionsExt/init.jl
@@ -18,7 +18,7 @@ function EvoTrees.init_core(params::EvoTrees.EvoTypes, device::Type{<:EvoTrees.G
     !isnothing(offset) && (pred .+= _to_device(backend, collect(offset')))
 
     ∇ = KernelAbstractions.zeros(backend, T, 2 * K + 1, nobs)
-    h∇ = KernelAbstractions.zeros(backend, Float64, 2 * K + 1, maximum(featbins), length(featbins), 2^params.max_depth - 1)
+    h∇ = KernelAbstractions.zeros(backend, Int64, 2 * K + 1, maximum(featbins), length(featbins), 2^params.max_depth - 1)
     @assert (size(y, ndims(y)) == length(w) && minimum(w) > 0)
     ∇[end, :] .= w
